requirements.txt
plotly
flask
matplotlib
gpiod
RPI.BME280